+++ src/routes/+page.server.ts
@@ -6,7 +6,17 @@ export const actions: Actions = {
 		const formData = await request.formData();
 		const theme = formData.get('theme')?.toString() ?? 'skeleton';
 		// Sets the selected theme to the cookie
-		cookies.set('theme', theme, { path: '/' });
+		cookies.set('theme', theme, {
+			path: '/',
+			httpOnly: true,
+			// only requests from same site can send cookies
+			// https://developer.mozilla.org/en-US/docs/Glossary/CSRF
+			sameSite: 'lax',
+			// only sent over HTTPS in production
+			secure: process.env.NODE_ENV === 'production',
+			// set cookie to expire after a month
+			maxAge: 60 * 60 * 24 * 30
+		});
 		return { theme };
 	}
 };
